new Rust illumos target, split from Solaris
The "illumos" target is officially in the Rust toolchain now.  It is
available in beta builds from the official Rust CI/CD system, and we
expect it to be available in the next stable release of the toolchain.

This is a minimal change to split off the illumos platform code from the
Solaris platform code, to allow building on native illumos Rust bits,
and to make future maintenance easier on these increasingly divergent
platforms.